インターフェース（serviceクラス）とImplクラスに分割および不要ファイルの削除
