+++ login/src/main/java/com/example/demo/service/ContactService.java
@@ -1,69 +1,14 @@
 package com.example.demo.service;
 
-import java.util.Optional;
-
-import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 
 import com.example.demo.entity.Contact;
 import com.example.demo.form.ContactsForm;
-import com.example.demo.repository.ContactRepository;
 
-@Service
-public class ContactService {
-
-	private final ContactRepository contactRepository;
-
-	public ContactService(ContactRepository contactRepository) {
-		this.contactRepository = contactRepository;
-	}
-
-	public String getContactDetail(Long id, Model model) {
-		Optional<Contact> contactOptional = contactRepository.findById(id);
-		if (contactOptional.isPresent()) {
-			Contact contact = contactOptional.get();
-			model.addAttribute("contact", contact);
-			return "contactsid";
-		} else {
-			return "error_page";
-		}
-	}
-
-	public String edit(Long id, Model model) {
-		Optional<Contact> contactOptional = contactRepository.findById(id);
-		if (contactOptional.isPresent()) {
-			Contact contact = contactOptional.get();
-			model.addAttribute("contact", contact);
-			return "contactsedit";
-		} else {
-			return "error";
-		}
-	}
-
-	public String update(Long id, ContactsForm form) {
-		Optional<Contact> optionalContact = contactRepository.findById(id);
-		Contact contact = optionalContact.orElseThrow(() -> new RuntimeException("お問い合わせが見つかりません: " + id));
-
-		contact.setLastName(form.getLastName());
-		contact.setFirstName(form.getFirstName());
-		contact.setEmail(form.getEmail());
-		contact.setPhone(form.getPhone());
-		contact.setZipCode(form.getZipCode());
-		contact.setAddress(form.getAddress());
-		contact.setBuildingName(form.getBuildingName());
-		contact.setContactType(form.getContactType());
-		contact.setBody(form.getBody());
-
-		contactRepository.save(contact);
-
-		return "redirect:/admin/contacts/" + id;
-	}
-
-	public Contact getById(Long id) {
-		return contactRepository.findById(id).orElse(null);
-	}
-
-	public void delete(Long id) {
-		contactRepository.deleteById(id);
-	}
+public interface ContactService {
+    String getContactDetail(Long id, Model model);
+    String edit(Long id, Model model);
+    String update(Long id, ContactsForm form);
+    Contact getById(Long id);
+    void delete(Long id);
 }
